adding react  sematic ui
