+++ package.json
@@ -7,14 +7,15 @@
     "@react-pdf/renderer": "^1.6.8",
     "axios": "^0.19.2",
     "js-yaml": "^3.13.1",
+    "minimist": "0.2.1",
     "react": "^16.12.0",
     "react-cv": "^1.1.0",
     "react-dom": "^16.12.0",
     "react-pdf": "^4.1.0",
     "react-router-dom": "^5.1.2",
     "react-undraw-illustrations": "^2.0.3",
-    "serialize-javascript": "^3.0.0",
-    "minimist": "0.2.1"
+    "semantic-ui-react": "^0.88.2",
+    "serialize-javascript": "^3.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.8.4",
@@ -25,12 +26,12 @@
     "babel-loader": "^8.0.2",
     "css-loader": "^1.0.0",
     "cypress": "^4.0.2",
+    "js-yaml": "3.13.1",
     "style-loader": "^0.23.0",
     "uglifyjs-webpack-plugin": "^2.2.0",
     "webpack": "^4.19.1",
     "webpack-cli": "^3.1.1",
-    "webpack-dev-server": "^3.1.8",
-    "js-yaml": "3.13.1"
+    "webpack-dev-server": "^3.1.8"
   },
   "scripts": {
     "start": "webpack-dev-server --inline --mode development --open",
