+++ src/components/Hello.jsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+const Hello = () => {
+  return (
+    <div className="ui large" style={{
+      background: 'url("/src/pic/IMG_1696.png")', 
+      backgroundSize: 'cover', height: '95vh',
+    }}>
+    </div>
+  );
+};
+
+export default Hello;
