+++ src/index.js
@@ -1,14 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import About from "./About";
-import Hello from "./Hello";
-import Header from "./Header";
-import Footer from "./Footer";
+import About from "./components/About";
+import Hello from "./components/Hello";
+import Header from "./components/Header";
+import Footer from "./components/Footer";
 import { Switch, Route, BrowserRouter } from "react-router-dom";
-import Projects from "./Projects";
-
-
-
+import Projects from "./components/Projects";
 
 const App = () => {
   return (
@@ -18,7 +15,6 @@ const App = () => {
         <Route exact path="/" component={Hello}></Route>
         <Route exact path="/about" component={About}></Route>
         <Route exact path="/projects" component={Projects}></Route>
-
       </Switch>
       <Footer />
     </>
